requirements.txt
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Markdown==3.0.1
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
